Remove `.dart` from `file_structure_strategy` dir
