+++ lib/generation/generators/value_objects/file_structure_strategy/bloc_file_structure_strategy.dart
@@ -1,4 +1,4 @@
-import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy.dart/pb_file_structure_strategy.dart';
+import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/pb_file_structure_strategy.dart';
 import 'package:parabeac_core/generation/generators/writers/pb_page_writer.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_project.dart';
