+++ lib/generation/generators/value_objects/file_structure_strategy/pb_file_structure_strategy.dart
