+++ lib/generation/generators/value_objects/file_structure_strategy/riverpod_file_structure_strategy.dart
@@ -1,16 +1,16 @@
 import 'dart:io';
 
+import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/pb_file_structure_strategy.dart';
 import 'package:parabeac_core/generation/generators/writers/pb_page_writer.dart';
-import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy.dart/pb_file_structure_strategy.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_project.dart';
 
-class ProviderFileStructureStrategy extends FileStructureStrategy {
-  final RELATIVE_PROVIDER_PATH = 'lib/providers/';
+class RiverpodFileStructureStrategy extends FileStructureStrategy {
+  final RELATIVE_PROVIDER_PATH = 'lib/riverpod/';
   final RELATIVE_MODEL_PATH = 'lib/models/';
   var _providersPath;
   var _modelsPath;
 
-  ProviderFileStructureStrategy(
+  RiverpodFileStructureStrategy(
       String genProjectPath, PBPageWriter pageWriter, PBProject pbProject)
       : super(genProjectPath, pageWriter, pbProject) {
     _providersPath = '${genProjectPath}${RELATIVE_PROVIDER_PATH}';
@@ -31,7 +31,7 @@ class ProviderFileStructureStrategy extends FileStructureStrategy {
     Directory(_modelsPath).createSync(recursive: true);
   }
 
-  void writeProviderModelFile(String code, String fileName) {
+  void writeRiverpodModelFile(String code, String fileName) {
     super
         .pageWriter
         .write(code, '${_modelsPath}${fileName}.dart'); // Removed .g
