+++ lib/generation/generators/value_objects/generation_configuration/riverpod_generation_configuration.dart
@@ -1,5 +1,5 @@
 import 'package:parabeac_core/generation/generators/middleware/state_management/riverpod_middleware.dart';
-import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy.dart/riverpod_file_structure_strategy.dart';
+import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/riverpod_file_structure_strategy.dart';
 import 'package:parabeac_core/generation/generators/value_objects/generation_configuration/pb_generation_configuration.dart';
 import 'package:parabeac_core/generation/generators/writers/pb_flutter_writer.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_project.dart';
