+++ lib/generation/generators/value_objects/generation_configuration/stateful_generation_configuration.dart
@@ -1,5 +1,5 @@
 import 'package:parabeac_core/generation/generators/middleware/state_management/stateful_middleware.dart';
-import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy.dart/flutter_file_structure_strategy.dart';
+import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/flutter_file_structure_strategy.dart';
 import 'package:parabeac_core/generation/generators/value_objects/generation_configuration/pb_generation_configuration.dart';
 
 class StatefulGenerationConfiguration extends GenerationConfiguration {
